Moved the IndexedLattice3d class into another file
