+++ src/form.rs
@@ -1,5 +1,4 @@
 /// Generic N-form in D dimensions
-/// See: https://blog.rust-lang.org/inside-rust/2021/09/06/Splitting-const-generics.html
 #[derive(Debug, Clone, Copy)]
 pub struct Form<T, const D: usize, const N: usize>([T; binomial_coeff_recursive(D, N)])
 where
@@ -16,11 +15,11 @@ where
 
 /// Implements the binomial coefficient n over k recursively.
 ///
-/// ```
+///
 ///  / n \   / n-1 \   / n-1 \      / n \   / n \
 ///  |   | = |     | + |     |  and |   | = |   | = 1 for all n;
 ///  \ k /   \ k-1 /   \  k  /      \ 0 /   \ n /
-/// ```
+///
 ///
 pub const fn binomial_coeff_recursive(n: usize, k: usize) -> usize {
     match (n, k) {
